refactor(taskflow): move shell + store composition to app
