+++ examples/taskflow/androidApp/src/main/kotlin/org/reduxkotlin/sample/taskflow/app/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import org.reduxkotlin.sample.taskflow.App
+import org.reduxkotlin.sample.taskflow.app.App
 import org.reduxkotlin.sample.taskflow.infra.platform.AndroidContextHolder
 
 class MainActivity : ComponentActivity() {
