+++ examples/taskflow/composeApp/src/iosMain/kotlin/MainViewController.kt
@@ -1,6 +1,7 @@
 package org.reduxkotlin.sample.taskflow
 
 import androidx.compose.ui.window.ComposeUIViewController
+import org.reduxkotlin.sample.taskflow.app.App
 
 // PascalCase is the Compose-MP convention for the iOS view-controller factory
 // (exported to Swift as `MainViewControllerKt.MainViewController()`).
