+++ examples/taskflow/composeApp/src/jvmMain/kotlin/main.kt
@@ -2,6 +2,7 @@ package org.reduxkotlin.sample.taskflow
 
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import org.reduxkotlin.sample.taskflow.app.App
 
 fun main() = application {
     Window(onCloseRequest = ::exitApplication, title = "TaskFlow") {
