+++ examples/taskflow/composeApp/src/wasmJsMain/kotlin/main.kt
@@ -3,6 +3,7 @@ package org.reduxkotlin.sample.taskflow
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
+import org.reduxkotlin.sample.taskflow.app.App
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
